Fixed could not start model with profiler bug

+++ csrc/mc2/mega_moe/op_kernel/arch35/mega_moe_combine_send.h
@@ -52,7 +52,7 @@ __aicore__ inline void CombineTokens(
         uint32_t topkIdx = tripleTensor.GetValue(tileIdx * 8 + 2);
         gmRemoteD.SetGlobalBuffer(reinterpret_cast<__gm__ ElementMMadOut2*>(
             GetRankWinAddrWithOffset(toRankId, gmRemoteBaseOffset)));
-        uint64_t gmDstOffset = (tokenIdx * params.tilingData->topK + topkIdx) * n + nLoc;
+        uint64_t gmDstOffset = (static_cast<uint64_t>(tokenIdx) * params.tilingData->topK + topkIdx) * n + nLoc;
         AscendC::DataCopyPad(gmRemoteD[gmDstOffset],
             l0cOutUbGMM2[tileIdx * Get<N_VALUE>(actualBlockShape)], ub2GmParams);
     }
@@ -435,7 +435,7 @@ __aicore__ inline void CombineQuantizedTokens(
     __gm__ void* dstPeermemPtr = GetRankWinAddrWithOffset(toRankId, gmRemoteOffset);
     gmRemoteD.SetGlobalBuffer(reinterpret_cast<__gm__ QuantOutType*>(dstPeermemPtr));
 
-    uint64_t dstBaseOffset = (tokenIdx * params.tilingData->topK + topkIdx) * quantTokenSize;
+    uint64_t dstBaseOffset = (static_cast<uint64_t>(tokenIdx) * params.tilingData->topK + topkIdx) * quantTokenSize;
 
     AscendC::DataCopyExtParams singleCopyParams{
         1, static_cast<uint32_t>(quantTokenSize), 0, 0, 0};
